+++ parser/src/processors/import.rs
@@ -8,7 +8,7 @@ impl super::Processor for Import {
         parser: &mut super::Parser,
         page_idx: usize,
         processed_page_idx: usize,
-        page_hash: u64,
+        page_hash: usize,
     ) -> bool {
         let duplicate = if self.reference == "" {
             false
